Address Copilot feedback
